Ajustes e tratamento de erros.

+++ at-ipoo_01-api_funcionario/src/main/java/org/example/service/FuncionarioService.java
@@ -26,22 +26,22 @@ public class FuncionarioService {
     }
 
     public Funcionario update(int id, Funcionario funcionarioAtualizado){
-        Optional<Funcionario> funcionarioExistente = funcionarioRepository.findById(id);
-        if (funcionarioExistente.isPresent()){
-            Funcionario funcionario = funcionarioExistente.get();
-            funcionario.setNome(funcionarioAtualizado.getNome());
-            funcionario.setTelefone(funcionarioAtualizado.getTelefone());
-            funcionario.setEmail(funcionarioAtualizado.getEmail());
-            funcionario.setEndereco(funcionarioAtualizado.getEndereco());
-            funcionario.setCidade(funcionarioAtualizado.getCidade());
-            funcionario.setSalario(funcionarioAtualizado.getSalario());
-            funcionario.setDataNasc(funcionarioAtualizado.getDataNasc());
-            funcionario.setDataCadas(funcionarioAtualizado.getDataCadas());
-            return funcionarioRepository.save(funcionario);
-        }
-        return null;
+        Funcionario funcionario = funcionarioRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Funcionário não encontrado com ID: " + id));
+        funcionario.setNome(funcionarioAtualizado.getNome());
+        funcionario.setTelefone(funcionarioAtualizado.getTelefone());
+        funcionario.setEmail(funcionarioAtualizado.getEmail());
+        funcionario.setEndereco(funcionarioAtualizado.getEndereco());
+        funcionario.setCidade(funcionarioAtualizado.getCidade());
+        funcionario.setSalario(funcionarioAtualizado.getSalario());
+        funcionario.setDataNasc(funcionarioAtualizado.getDataNasc());
+        funcionario.setDataCadas(funcionarioAtualizado.getDataCadas());
+        return funcionarioRepository.save(funcionario);
     }
     public void delete(int id){
+        if (!funcionarioRepository.existsById(id)) {
+            throw new RuntimeException("Funcionário não encontrado com ID: " + id);
+        }
         funcionarioRepository.deleteById(id);
     }
 }
